Is animal in ROI?
Updated program to detect if an animal is in a user defined ROI. It can
do this based on the centroid, the ellipsis described around the
animal, or the animal itself.

The main issue is tracking the animal’s head specifically. Once we can
do this, the rest of the process will be trivial.