Tracking works!
Big change to IRtrack. Details are commented in the code — but it
allows the user to track the animal based on the location of the head.
It also allows for the drawing of a polygonal ROI and when the animal
is detected in said ROI, a function can be triggered.